# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Courtroom AI Simulator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type RoleController =
  (typeof RoleController)[keyof typeof RoleController];

export const RoleController = {
  user: "user",
  ai: "ai",
} as const;

export type CourtPhase = (typeof CourtPhase)[keyof typeof CourtPhase];

export const CourtPhase = {
  opening_statements: "opening_statements",
  prosecution_case: "prosecution_case",
  defense_case: "defense_case",
  closing_arguments: "closing_arguments",
  verdict: "verdict",
  concluded: "concluded",
} as const;

export interface RoleAssignment {
  judge: RoleController;
  prosecutor: RoleController;
  defense: RoleController;
}

export type TranscriptEntryRole =
  (typeof TranscriptEntryRole)[keyof typeof TranscriptEntryRole];

export const TranscriptEntryRole = {
  judge: "judge",
  prosecutor: "prosecutor",
  defense: "defense",
  system: "system",
} as const;

export type TranscriptEntryControlledBy =
  (typeof TranscriptEntryControlledBy)[keyof typeof TranscriptEntryControlledBy];

export const TranscriptEntryControlledBy = {
  user: "user",
  ai: "ai",
  system: "system",
} as const;

export interface TranscriptEntry {
  id: string;
  role: TranscriptEntryRole;
  speaker: string;
  content: string;
  timestamp: string;
  controlledBy: TranscriptEntryControlledBy;
}

export interface Development {
  id: string;
  title: string;
  content: string;
  timestamp: string;
}

export interface CaseSession {
  caseId: string;
  title: string;
  caseText: string;
  phase: CourtPhase;
  roles: RoleAssignment;
  transcript: TranscriptEntry[];
  developments: Development[];
  createdAt: string;
  updatedAt: string;
}

export type CaseListResponseCasesItem = {
  caseId: string;
  title: string;
  phase: CourtPhase;
  createdAt: string;
};

export interface CaseListResponse {
  cases: CaseListResponseCasesItem[];
}

export interface CreateCaseRequest {
  title: string;
  caseText: string;
  roles: RoleAssignment;
}

export interface UpdateRolesRequest {
  roles: RoleAssignment;
}

export type SpeakRequestRole =
  (typeof SpeakRequestRole)[keyof typeof SpeakRequestRole];

export const SpeakRequestRole = {
  judge: "judge",
  prosecutor: "prosecutor",
  defense: "defense",
} as const;

export interface SpeakRequest {
  role: SpeakRequestRole;
  content: string;
  triggerAiResponses?: boolean;
}

export interface SpeakResponse {
  userEntry: TranscriptEntry;
  aiResponses: TranscriptEntry[];
  updatedTranscript: TranscriptEntry[];
}

export type AiTurnRequestRole =
  (typeof AiTurnRequestRole)[keyof typeof AiTurnRequestRole];

export const AiTurnRequestRole = {
  judge: "judge",
  prosecutor: "prosecutor",
  defense: "defense",
} as const;

export interface AiTurnRequest {
  role: AiTurnRequestRole;
  context?: string;
}

export interface AiTurnResponse {
  entry: TranscriptEntry;
  updatedTranscript: TranscriptEntry[];
}

export interface AutoProceedResponse {
  entries: TranscriptEntry[];
  updatedTranscript: TranscriptEntry[];
}

export interface AddDevelopmentRequest {
  title: string;
  content: string;
}

export interface UpdatePhaseRequest {
  phase: CourtPhase;
}