# lib/api-zod/src/generated/types/caseListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Courtroom AI Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { CaseListResponseCasesItem } from "./caseListResponseCasesItem";

export interface CaseListResponse {
  cases: CaseListResponseCasesItem[];
}