# lib/api-zod/src/generated/types/courtPhase.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Courtroom AI Simulator API
 * OpenAPI spec version: 0.1.0
 */

export type CourtPhase = (typeof CourtPhase)[keyof typeof CourtPhase];

export const CourtPhase = {
  opening_statements: "opening_statements",
  prosecution_case: "prosecution_case",
  defense_case: "defense_case",
  closing_arguments: "closing_arguments",
  verdict: "verdict",
  concluded: "concluded",
} as const;