# lib/api-zod/src/generated/types/createCaseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Courtroom AI Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { RoleAssignment } from "./roleAssignment";

export interface CreateCaseRequest {
  title: string;
  caseText: string;
  roles: RoleAssignment;
}