# lib/api-zod/src/generated/types/development.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Courtroom AI Simulator API
 * OpenAPI spec version: 0.1.0
 */

export interface Development {
  id: string;
  title: string;
  content: string;
  timestamp: string;
}