# lib/api-zod/src/generated/types/roleAssignment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Courtroom AI Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { RoleController } from "./roleController";

export interface RoleAssignment {
  judge: RoleController;
  prosecutor: RoleController;
  defense: RoleController;
}