# lib/api-zod/src/generated/types/roleController.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Courtroom AI Simulator API
 * OpenAPI spec version: 0.1.0
 */

export type RoleController =
  (typeof RoleController)[keyof typeof RoleController];

export const RoleController = {
  user: "user",
  ai: "ai",
} as const;