# lib/api-zod/src/generated/types/transcriptEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Courtroom AI Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { TranscriptEntryControlledBy } from "./transcriptEntryControlledBy";
import type { TranscriptEntryRole } from "./transcriptEntryRole";

export interface TranscriptEntry {
  id: string;
  role: TranscriptEntryRole;
  speaker: string;
  content: string;
  timestamp: string;
  controlledBy: TranscriptEntryControlledBy;
}